+++ LoginUI/Data.swift
@@ -1,0 +1,14 @@
+//
+//  Data.swift
+//  LoginUI
+//
+//  Created by Taqtile on 16/05/18.
+//  Copyright © 2018 Taqtile. All rights reserved.
+//
+
+import Foundation
+
+struct Data: Codable {
+    let user: User?
+    let token: String?
+}
