+++ LoginUI/FormValidator.swift
@@ -1,0 +1,33 @@
+//
+//  FormValidator.swift
+//  LoginUI
+//
+//  Created by Taqtile on 16/05/18.
+//  Copyright © 2018 Taqtile. All rights reserved.
+//
+
+import Foundation
+
+class FormValidator {
+    func isValidEmail(testStr: String) -> Bool {
+        let emailRegEx = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}"
+        
+        let emailTest = NSPredicate(format:"SELF MATCHES %@", emailRegEx)
+        return emailTest.evaluate(with: testStr)
+    }
+    
+    func isValidPassword(testStr: String) -> Bool {
+        if testStr.count < 4 {return false}
+        return true
+    }
+    
+    func containOnlyLetters(testStr: String) -> Bool {
+        for chr in testStr {
+            if (!(chr >= "a" && chr <= "z") && !(chr >= "A" && chr <= "Z") ) {
+                return false
+            }
+        }
+        return true
+    }
+    
+}
