+++ LoginUI/Login.swift
@@ -1,0 +1,14 @@
+//
+//  Login.swift
+//  LoginUI
+//
+//  Created by Taqtile on 16/05/18.
+//  Copyright © 2018 Taqtile. All rights reserved.
+//
+
+import Foundation
+
+struct Login: Codable {
+    let data: Data?
+    let errors: [Error]?
+}
