+++ LoginUI/Profile.swift
@@ -1,0 +1,13 @@
+//
+//  Profile.swift
+//  LoginUI
+//
+//  Created by Taqtile on 16/05/18.
+//  Copyright © 2018 Taqtile. All rights reserved.
+//
+
+import Foundation
+
+struct Profile: Codable {
+    let data: User?
+}
